add support for multiple source files; fix incorrect Ast alignment

+++ src/literals.rs
@@ -1,12 +1,12 @@
 use core::num::ParseIntError;
 use std::num::ParseFloatError;
 
-pub fn parse_int_lit(text: &str) -> Result<i128, ParseIntError> {
+pub fn parse_int_lit(text: &str) -> Result<i64, ParseIntError> {
     match text.get(..2) {
-        Some("0b") => i128::from_str_radix(&text[2..], 2),
-        Some("0o") => i128::from_str_radix(&text[2..], 8),
-        Some("0x") => i128::from_str_radix(&text[2..], 16),
-        _ => i128::from_str_radix(text, 10),
+        Some("0b") => i64::from_str_radix(&text[2..], 2),
+        Some("0o") => i64::from_str_radix(&text[2..], 8),
+        Some("0x") => i64::from_str_radix(&text[2..], 16),
+        _ => i64::from_str_radix(text, 10),
     }
 }
